Build fixes.

+++ src/ColumnListView.cpp
@@ -518,7 +518,7 @@ BRow::BRow(float height)
 BRow::~BRow()
 {
 	while (true) {
-		BField *field = (BField*) fFields.RemoveItem(0);
+		BField *field = (BField*) fFields.RemoveItem((long int)0);
 		if (field == 0)
 			break;
 		
@@ -927,7 +927,7 @@ BColumnListView::BColumnListView(BRect rect, const char *name, uint32 resizingMo
 BColumnListView::~BColumnListView()
 {
 	while (true) {
-		BColumn *column = (BColumn*) fColumns.RemoveItem(0);
+		BColumn *column = (BColumn*) fColumns.RemoveItem((long int)0);
 		if (column == 0)
 			break;
